+++ cypress/e2e/integration/ieee8021x/create.spec.ts
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  **********************************************************************/
 
-import { AuthenticationProtocols, Config } from '../../../../src/app/ieee8021x/ieee8021x.constants'
+import { AuthenticationProtocols } from '../../../../src/app/ieee8021x/ieee8021x.constants'
 import { httpCodes } from '../../fixtures/api/httpCodes'
 import { wiredConfigs, wirelessConfigs } from '../../fixtures/formEntry/ieee8021x'
 import * as api8021x from '../../fixtures/api/ieee8021x'
 import { noConfigsResponse } from '../../fixtures/api/ieee8021x'
+import { IEEE8021xConfig } from 'src/models/models'
 
 beforeEach('clear cache and login', () => {
   cy.setup()
@@ -17,7 +18,7 @@ beforeEach('clear cache and login', () => {
 })
 
 describe('test ieee8021x wired config creation', () => {
-  const createdConfigs: Config[] = []
+  const createdConfigs: IEEE8021xConfig[] = []
   wiredConfigs.forEach((config) => {
     it(`should create config ${config.profileName}`, () => {
       createdConfigs.push(config)
@@ -32,13 +33,15 @@ describe('test ieee8021x wired config creation', () => {
       cy.wait('@getAll')
       // Check that the ieee8021x config was successful
       cy.get('mat-cell').contains(config.profileName)
-      cy.get('mat-cell').contains(AuthenticationProtocols.labelForValue(config.authenticationProtocol))
+      cy.get('mat-cell').contains(
+        AuthenticationProtocols.find((x) => x.value === config.authenticationProtocol)?.label ?? 'error'
+      )
     })
   })
 })
 
 describe('test ieee8021x wireless config creation', () => {
-  const createdConfigs: Config[] = []
+  const createdConfigs: IEEE8021xConfig[] = []
   wirelessConfigs.forEach((config) => {
     it(`should create config ${config.profileName}`, () => {
       createdConfigs.push(config)
@@ -53,7 +56,9 @@ describe('test ieee8021x wireless config creation', () => {
       cy.wait('@getAll')
       // Check that the ieee8021x config was successful
       cy.get('mat-cell').contains(config.profileName)
-      cy.get('mat-cell').contains(AuthenticationProtocols.labelForValue(config.authenticationProtocol))
+      cy.get('mat-cell').contains(
+        AuthenticationProtocols.find((x) => x.value === config.authenticationProtocol)?.label ?? 'error'
+      )
     })
   })
 })