+++ src/app/configs/configs.service.ts
@@ -8,7 +8,7 @@ import { Injectable } from '@angular/core'
 import { Observable, throwError } from 'rxjs'
 import { catchError } from 'rxjs/operators'
 import { environment } from 'src/environments/environment'
-import { CIRAConfig, CIRAConfigResponse, PageEventOptions } from 'src/models/models'
+import { CIRAConfig, DataWithCount, PageEventOptions } from 'src/models/models'
 import { AuthService } from '../auth.service'
 
 @Injectable({
@@ -22,14 +22,14 @@ export class ConfigsService {
     private readonly http: HttpClient
   ) {}
 
-  getData(pageEvent?: PageEventOptions): Observable<CIRAConfigResponse> {
+  getData(pageEvent?: PageEventOptions): Observable<DataWithCount<CIRAConfig>> {
     let query = this.url
     if (pageEvent) {
       query += `?$top=${pageEvent.pageSize}&$skip=${pageEvent.startsFrom}&$count=${pageEvent.count}`
     } else {
       query += '?$count=true'
     }
-    return this.http.get<CIRAConfigResponse>(query).pipe(
+    return this.http.get<DataWithCount<CIRAConfig>>(query).pipe(
       catchError((err) => {
         const errorMessages = this.authService.onError(err)
         return throwError(errorMessages)