+++ src/app/devices/audit-log/audit-log.component.spec.ts
@@ -4,7 +4,7 @@
  **********************************************************************/
 
 import { ComponentFixture, TestBed } from '@angular/core/testing'
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations'
+import { BrowserAnimationsModule, NoopAnimationsModule } from '@angular/platform-browser/animations'
 import { ActivatedRoute, RouterModule } from '@angular/router'
 import { of } from 'rxjs'
 import { DevicesService } from '../devices.service'
@@ -22,7 +22,7 @@ describe('AuditLogComponent', () => {
 
     await TestBed.configureTestingModule({
       imports: [
-        BrowserAnimationsModule,
+        NoopAnimationsModule,
         RouterModule,
         AuditLogComponent
       ],
@@ -42,10 +42,6 @@ describe('AuditLogComponent', () => {
     fixture = TestBed.createComponent(AuditLogComponent)
     component = fixture.componentInstance
     fixture.detectChanges()
-  })
-
-  afterEach(() => {
-    TestBed.resetTestingModule()
   })
 
   afterEach(() => {