+++ src/app/ieee8021x/ieee8021x.component.ts
@@ -8,12 +8,12 @@ import { MatDialog } from '@angular/material/dialog'
 import { MatSnackBar } from '@angular/material/snack-bar'
 import { Router } from '@angular/router'
 import { finalize } from 'rxjs/operators'
-import { PageEventOptions } from 'src/models/models'
+import { DataWithCount, IEEE8021xConfig, PageEventOptions } from 'src/models/models'
 import { AreYouSureDialogComponent } from '../shared/are-you-sure/are-you-sure.component'
 import SnackbarDefaults from '../shared/config/snackBarDefault'
 import { IEEE8021xService } from './ieee8021x.service'
 import { MatPaginator, PageEvent } from '@angular/material/paginator'
-import { AuthenticationProtocols, ConfigsResponse } from './ieee8021x.constants'
+import { AuthenticationProtocols } from './ieee8021x.constants'
 import {
   MatTable,
   MatColumnDef,
@@ -31,6 +31,7 @@ import { MatProgressBar } from '@angular/material/progress-bar'
 import { MatIcon } from '@angular/material/icon'
 import { MatButton, MatIconButton } from '@angular/material/button'
 import { MatToolbar } from '@angular/material/toolbar'
+import { ToolkitPipe } from '../shared/pipes/toolkit.pipe'
 
 @Component({
   selector: 'app-ieee8021x',
@@ -38,6 +39,7 @@ import { MatToolbar } from '@angular/material/toolbar'
   styleUrls: ['./ieee8021x.component.scss'],
   standalone: true,
   imports: [
+    ToolkitPipe,
     MatToolbar,
     MatButton,
     MatIcon,
@@ -59,7 +61,7 @@ import { MatToolbar } from '@angular/material/toolbar'
   ]
 })
 export class IEEE8021xComponent implements OnInit {
-  pagedConfigs: ConfigsResponse = {
+  pagedConfigs: DataWithCount<IEEE8021xConfig> = {
     data: [],
     totalCount: 0
   }
@@ -100,7 +102,7 @@ export class IEEE8021xComponent implements OnInit {
         })
       )
       .subscribe({
-        next: (pagedConfigs: ConfigsResponse) => {
+        next: (pagedConfigs) => {
           this.pagedConfigs = pagedConfigs
         },
         error: () => {